+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Fieldset captions are gone (https://github.com/edvin/tornadofx/issues/399)
 - AutoCompleteComboBoxSkin no longer throws upon reconfiguration
 
+### Changes
+
+- AutoCompleteComboBoxSkin: Added an option to use automatic width for the popup
+
 ## [1.7.9] - 2017-08-04
 
 ### Additions